Module Make.Auxiliary

Functor to create auxiliary functions used by Abstract and Blockchain Protocols.

Parameters

module Event : Simulator.Events.Event

module Queue : Simulator.Events.EventQueue with type ev = Event.t

module Logger : Simulator.Logging.Logger with type ev = Event.t

module Network : Abstractions.Network.Network

Signature

val create_nodes : (int -> Abstractions.Network.links -> int -> 'a) -> (int, 'a) Stdlib.Hashtbl.t

Create the nodes to be used in the simulation, and produce the node and link creation JSON logs.

parameter init
function to create the initial state of a node

val add_initial_events : Queue.ev list -> unit

Add the initial events that jumpstart the simulation loop to the Event Queue.

parameter events
the list of events to be added to the queue