Parameter Blockchain.10-Network

type msg

The type of the messages being exchanged

module MessagesExchanged : Simulator.Statistics.Stats

Stats module to keep track of the number of messages exchanged during the simulation

module MegabytesExchanged : Simulator.Statistics.Stats

Stats module to keep track of the total megabytes exchanged during the simulation

val send : int -> int -> msg -> unit

Primitive for point-to-point message exchange.

parameter sender_node_id
the id of the node sending the message

parameter target_node_id
the id of the node receiving the message

parameter message
the contents of the message being sent

val send_to_neighbors : int -> msg -> unit

Primitive to send a message to all neighbors of a given node

parameter sender_node_id
the id of the node sending the message

parameter message
the contents of the message being sent

val gossip : int -> msg -> unit

Primitive to abstract the gossip protocol. Gossips a message to every node in the network. Used for slightly more relaxed scenarios, since it assumes that limited-bandwidth=false.

parameter sender_node_id
the id of the node sending the message

parameter message
the contents of the message being sent

val get_neighbours : int -> Abstractions.Network.links

Primitive to obtain the neighbors of a given node.

parameter node_id
the id of the node whose neighbors we want

val clear : unit -> unit

Primitive to reset all data stored by the network.

val get_links : unit -> Abstractions.Network.network_links

Primitive to get the links assigned to each node in the network.

val get_regions : unit -> Abstractions.Network.regions

Primitive to get the region assigned to each node in the network.